+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- Add functionality to backup CAPI clusters.
+
 ## [3.1.0] - 2022-06-15
 
 ### Added